requirements.txt
openai>=1.12.0
playwright>=1.40.0
pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.0
aiofiles>=23.0.0 